+++ epics/refresh.js
@@ -4,9 +4,20 @@ export function refreshEpic(resource) {
   return (action$) => action$
     .ofType('REFRESH')
     .filter(action => {
-      const { name, path } = action.meta;
-      const resPath = resource.optionsTemplate.path || '';
-      return resource.isVisible() && resPath.startsWith(path);
+      const { path } = action.meta;
+      const options = resource.optionsTemplate;
+      const resPath = options.path || (options.GET || {}).path || '';
+
+      if (!resource.isVisible()) return false;
+
+      let refresh = resPath.startsWith(path);
+
+      if (options.shouldRefresh) {
+        refresh = options.shouldRefresh(resource, action, refresh);
+      }
+
+      return refresh;
+
     })
     .debounceTime(100)
     .map(action => {